+++ app/services/activity/import.rb
@@ -6,7 +6,10 @@ class Activity
       end
 
       def csv_column
-        return "Implementing organisation names" if column == :implementing_organisation_id
+        if [:implementing_organisation_id, :implementing_org_participations].include?(column)
+          return "Implementing organisation names"
+        end
+
         ACTIVITY_CSV_COLUMNS.dig(column, :heading) || column.to_s
       end
     }
@@ -607,7 +610,7 @@ class Activity
       end
 
       def convert_ispf_oda_partner_countries(ispf_oda_partner_countries)
-        valid_codes = ispf_partner_country_options(oda: true).map { |country| country.code.to_s }
+        valid_codes = ispf_partner_countries_options(oda: true).map { |country| country.code.to_s }
 
         ispf_oda_partner_countries.split("|").map do |code|
           unless valid_codes.include?(code)
@@ -619,7 +622,7 @@ class Activity
       end
 
       def convert_ispf_non_oda_partner_countries(ispf_non_oda_partner_countries)
-        valid_codes = ispf_partner_country_options(oda: false).map { |country| country.code.to_s }
+        valid_codes = ispf_partner_countries_options(oda: false).map { |country| country.code.to_s }
 
         ispf_non_oda_partner_countries.split("|").map do |code|
           unless valid_codes.include?(code)