+++ db/migrate/20230116171600_add_export_filename_to_report.rb
@@ -1,0 +1,5 @@
+class AddExportFilenameToReport < ActiveRecord::Migration[6.1]
+  def change
+    add_column :reports, :export_filename, :string
+  end
+end