+++ spec/services/export/s3_uploader_config_spec.rb
@@ -2,6 +2,8 @@ require "rails_helper"
 
 module Export
   RSpec.describe S3UploaderConfig do
+    subject(:config) { S3UploaderConfig.new(use_public_bucket: true) }
+
     context "when an expected credential is missing from VCAP_SERVICES" do
       around(:each) do |example|
         vcap_services = <<~JSON
@@ -22,7 +24,7 @@ module Export
       end
 
       it "raises a helpful error message" do
-        expect { S3UploaderConfig.region }.to raise_error(KeyError, /key not found: "aws_region"/)
+        expect { config.region }.to raise_error(KeyError, /key not found: "aws_region"/)
       end
     end
 
@@ -46,7 +48,7 @@ module Export
       end
 
       it "raises a helpful error message" do
-        expect { S3UploaderConfig.region }.to raise_error(/AWS S3 credentials not found/)
+        expect { config.region }.to raise_error(/AWS S3 credentials not found/)
       end
     end
 
@@ -70,7 +72,7 @@ module Export
       end
 
       it "raises a helpful error message" do
-        expect { S3UploaderConfig.region }.to raise_error(/AWS S3 credentials not found/)
+        expect { config.region }.to raise_error(/AWS S3 credentials not found/)
       end
     end
 
@@ -85,7 +87,7 @@ module Export
       end
 
       it "raises a helpful error message" do
-        expect { S3UploaderConfig.region }.to raise_error(/AWS S3 credentials not found/)
+        expect { config.region }.to raise_error(/AWS S3 credentials not found/)
       end
     end
 
@@ -95,9 +97,18 @@ module Export
           {
             "aws-s3-bucket":[
                 {
+                   "name": "beis-roda-staging-s3-export-download-bucket-private",
+                   "credentials":{
+                      "bucket_name":"private_exports_bucket",
+                      "aws_access_key_id":"KEY_ID",
+                      "aws_secret_access_key":"SECRET_KEY",
+                      "aws_region":"eu-west-2"
+                   }
+                },
+                {
                    "name": "beis-roda-staging-s3-export-download-bucket",
                    "credentials":{
-                      "bucket_name":"exports_bucket",
+                      "bucket_name":"public_exports_bucket",
                       "aws_access_key_id":"KEY_ID",
                       "aws_secret_access_key":"SECRET_KEY",
                       "aws_region":"eu-west-2"
@@ -109,20 +120,34 @@ module Export
         ClimateControl.modify(VCAP_SERVICES: vcap_services) { example.run }
       end
 
-      it "returns the bucket_name" do
-        expect(S3UploaderConfig.bucket).to eq("exports_bucket")
-      end
-
       it "returns the key_id" do
-        expect(S3UploaderConfig.key_id).to eq("KEY_ID")
+        expect(config.key_id).to eq("KEY_ID")
       end
 
       it "returns the secret_key" do
-        expect(S3UploaderConfig.secret_key).to eq("SECRET_KEY")
+        expect(config.secret_key).to eq("SECRET_KEY")
       end
 
       it "returns the region" do
-        expect(S3UploaderConfig.region).to eq("eu-west-2")
+        expect(config.region).to eq("eu-west-2")
+      end
+
+      describe "#bucket" do
+        context "when initialized with `use_public_bucket: true`" do
+          subject(:config) { S3UploaderConfig.new(use_public_bucket: true) }
+
+          it "uses the public S3 bucket" do
+            expect(config.bucket).to eq("public_exports_bucket")
+          end
+        end
+
+        context "when initialized with `use_public_bucket: false`" do
+          subject(:config) { S3UploaderConfig.new(use_public_bucket: false) }
+
+          it "uses the private S3 bucket" do
+            expect(config.bucket).to eq("private_exports_bucket")
+          end
+        end
       end
     end
   end